fix(auth): use navigation.dispatch(reset) in AppLayout to bypass NativeTabs on iOS
SignOutGuard in root _layout.tsx used router.replace('/auth') which routes through
the focused navigator chain. On iOS with NativeTabs, this is silently dropped because
'auth' is not a tab route.

Replace with SignOutNavigator rendered inside AppLayout ((app) screen). useNavigation()
there returns the ROOT Stack's navigation prop — dispatching CommonActions.reset directly
to the root, bypassing NativeTabs entirely. Works on both iOS and Android.
