+++ apps/expo/app/_layout.tsx
@@ -2,20 +2,18 @@ import '../polyfills';
 
 import { ThemeProvider as NavThemeProvider } from '@react-navigation/native';
 import 'expo-dev-client';
-import { type Href, router, Stack } from 'expo-router';
+import { Stack } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 import '../global.css';
 
 import { clientEnvs } from '@packrat/env/expo-client';
 import { Alert, type AlertMethods } from '@packrat/ui/nativewindui';
 import * as Sentry from '@sentry/react-native';
-import { signOutRequestedAtom } from 'expo-app/features/auth/atoms/authAtoms';
 import { userStore } from 'expo-app/features/auth/store';
 import { useColorScheme, useInitialAndroidBarSync } from 'expo-app/lib/hooks/useColorScheme';
 import { Providers } from 'expo-app/providers';
 import { NAV_THEME } from 'expo-app/theme';
-import { useAtomValue, useSetAtom } from 'jotai';
-import { useEffect, useRef } from 'react';
+import { useRef } from 'react';
 
 Sentry.init({
   dsn: clientEnvs.EXPO_PUBLIC_SENTRY_DSN,
@@ -36,27 +34,6 @@ export {
 
 export let appAlert: React.RefObject<AlertMethods | null>;
 
-// Watches signOutRequestedAtom from OUTSIDE the navigation Stack so that router.replace
-// dispatches directly to the root navigator — not to NativeTabs' inner navigator,
-// which would silently drop the action on iOS because 'auth' isn't a tab route.
-// Using a dedicated atom (set only by signOut()) avoids false-positives during login
-// when tokenAtom transitions null → value.
-// Rendered inside <Providers> for Jotai context but before <NavThemeProvider><Stack>.
-function SignOutGuard() {
-  const signOutRequested = useAtomValue(signOutRequestedAtom);
-  const setSignOutRequested = useSetAtom(signOutRequestedAtom);
-
-  useEffect(() => {
-    if (signOutRequested) {
-      setSignOutRequested(false);
-      // safe-cast: '/auth' is a valid root route; Expo Router's Href accepts string paths directly.
-      router.replace('/auth' as Href);
-    }
-  }, [signOutRequested, setSignOutRequested]);
-
-  return null;
-}
-
 function RootLayout() {
   useInitialAndroidBarSync();
 
@@ -66,7 +43,6 @@ function RootLayout() {
 
   return (
     <Providers>
-      <SignOutGuard />
       <StatusBar
         key={`root-status-bar-${isDarkColorScheme ? 'light' : 'dark'}`}
         style={isDarkColorScheme ? 'light' : 'dark'}
